+++ isolates/src/main/resources/META-INF/native-image/main/jni-config.json
@@ -1,11 +1,73 @@
 [
 {
-  "name":"com.salesforce.apollo.domain.Demesne",
+  "name":"com.salesforce.apollo.domain.Smokin",
   "methods":[{"name":"main","parameterTypes":["java.lang.String[]"] }]
+},
+{
+  "name":"io.netty.channel.ChannelException"
+},
+{
+  "name":"io.netty.channel.DefaultFileRegion",
+  "fields":[
+    {"name":"file"}, 
+    {"name":"transferred"}
+  ]
+},
+{
+  "name":"io.netty.channel.kqueue.BsdSocket"
+},
+{
+  "name":"io.netty.channel.kqueue.KQueueEventArray"
+},
+{
+  "name":"io.netty.channel.kqueue.KQueueStaticallyReferencedJniMethods"
+},
+{
+  "name":"io.netty.channel.kqueue.Native"
+},
+{
+  "name":"io.netty.channel.unix.Buffer"
+},
+{
+  "name":"io.netty.channel.unix.DatagramSocketAddress",
+  "methods":[{"name":"<init>","parameterTypes":["byte[]","int","int","int","io.netty.channel.unix.DatagramSocketAddress"] }]
+},
+{
+  "name":"io.netty.channel.unix.DomainDatagramSocketAddress",
+  "methods":[{"name":"<init>","parameterTypes":["byte[]","int","io.netty.channel.unix.DomainDatagramSocketAddress"] }]
+},
+{
+  "name":"io.netty.channel.unix.ErrorsStaticallyReferencedJniMethods"
+},
+{
+  "name":"io.netty.channel.unix.FileDescriptor"
+},
+{
+  "name":"io.netty.channel.unix.LimitsStaticallyReferencedJniMethods"
+},
+{
+  "name":"io.netty.channel.unix.PeerCredentials",
+  "methods":[{"name":"<init>","parameterTypes":["int","int","int[]"] }]
+},
+{
+  "name":"io.netty.channel.unix.Socket"
+},
+{
+  "name":"java.io.FileDescriptor",
+  "fields":[{"name":"fd"}]
+},
+{
+  "name":"java.io.IOException"
 },
 {
   "name":"java.lang.Boolean",
   "methods":[{"name":"getBoolean","parameterTypes":["java.lang.String"] }]
+},
+{
+  "name":"java.lang.OutOfMemoryError"
+},
+{
+  "name":"java.lang.RuntimeException"
 },
 {
   "name":"java.lang.String",
@@ -20,5 +82,47 @@
     {"name":"getProperty","parameterTypes":["java.lang.String"] }, 
     {"name":"setProperty","parameterTypes":["java.lang.String","java.lang.String"] }
   ]
+},
+{
+  "name":"java.net.InetSocketAddress",
+  "methods":[{"name":"<init>","parameterTypes":["java.lang.String","int"] }]
+},
+{
+  "name":"java.net.PortUnreachableException"
+},
+{
+  "name":"java.nio.Buffer",
+  "fields":[
+    {"name":"limit"}, 
+    {"name":"position"}
+  ],
+  "methods":[
+    {"name":"limit","parameterTypes":[] }, 
+    {"name":"position","parameterTypes":[] }
+  ]
+},
+{
+  "name":"java.nio.DirectByteBuffer"
+},
+{
+  "name":"java.nio.channels.ClosedChannelException",
+  "methods":[{"name":"<init>","parameterTypes":[] }]
+},
+{
+  "name":"sun.management.VMManagementImpl",
+  "fields":[
+    {"name":"compTimeMonitoringSupport"}, 
+    {"name":"currentThreadCpuTimeSupport"}, 
+    {"name":"objectMonitorUsageSupport"}, 
+    {"name":"otherThreadCpuTimeSupport"}, 
+    {"name":"remoteDiagnosticCommandsSupport"}, 
+    {"name":"synchronizerUsageSupport"}, 
+    {"name":"threadAllocatedMemorySupport"}, 
+    {"name":"threadContentionMonitoringSupport"}
+  ]
+},
+{
+  "name":"sun.nio.ch.FileChannelImpl",
+  "fields":[{"name":"fd"}]
 }
 ]
