+++ protocols/src/main/java/com/salesforce/apollo/comm/grpc/DomainSocketServerInterceptor.java
@@ -15,17 +15,7 @@ import io.grpc.ServerCall;
 import io.grpc.ServerCallHandler;
 import io.grpc.ServerInterceptor;
 import io.grpc.Status;
-import io.netty.channel.Channel;
-import io.netty.channel.EventLoopGroup;
-import io.netty.channel.epoll.EpollDomainSocketChannel;
-import io.netty.channel.epoll.EpollEventLoopGroup;
-import io.netty.channel.epoll.EpollServerDomainSocketChannel;
-import io.netty.channel.kqueue.KQueue;
-import io.netty.channel.kqueue.KQueueDomainSocketChannel;
-import io.netty.channel.kqueue.KQueueEventLoopGroup;
-import io.netty.channel.kqueue.KQueueServerDomainSocketChannel;
 import io.netty.channel.unix.PeerCredentials;
-import io.netty.channel.unix.ServerDomainSocketChannel;
 
 /**
  * @author hal.hildebrand
@@ -34,18 +24,6 @@ import io.netty.channel.unix.ServerDomainSocketChannel;
 public class DomainSocketServerInterceptor implements ServerInterceptor {
 
     public static final Context.Key<PeerCredentials> PEER_CREDENTIALS_CONTEXT_KEY = Context.key("com.salesforce.apollo.PEER_CREDENTIALS");
-
-    public static Class<? extends Channel> getChannelType() {
-        return KQueue.isAvailable() ? KQueueDomainSocketChannel.class : EpollDomainSocketChannel.class;
-    }
-
-    public static EventLoopGroup getEventLoopGroup() {
-        return KQueue.isAvailable() ? new KQueueEventLoopGroup() : new EpollEventLoopGroup();
-    }
-
-    public static Class<? extends ServerDomainSocketChannel> getServerDomainSocketChannelClass() {
-        return KQueue.isAvailable() ? KQueueServerDomainSocketChannel.class : EpollServerDomainSocketChannel.class;
-    }
 
     @Override
     public <ReqT, RespT> ServerCall.Listener<ReqT> interceptCall(ServerCall<ReqT, RespT> call,
